move modal to NewDocumentView
