+++ app/frontend/src/views/HomeView.vue
@@ -21,8 +21,7 @@ const router = useRouter()
 const showModal = ref(false)
 const newName = ref('')
 const selectedBoardID = ref('')
-const showNewBoardModal = ref(false) 
-const newBoardName = ref('')
+
 
 const { get, data } = useGet<GetWhiteboardsResponse>(WhiteboardEndpoint)
 const client = useAPIClient()
@@ -37,20 +36,6 @@ async function saveNewName(boardId: string) {
   }
 }
 
-async function createNewBoard() {
-  try {
-    
-    const response = await client.post(WhiteboardEndpoint, { title: newBoardName.value })
-    const newBoardId = response.data.whiteBoardId 
-    get()
-    closeNewBoardModal()
-    router.push({ path: `d/${newBoardId}` }) 
-  } catch (error) {
-    console.error('Error al crear nueva pizarra:', error)
-  }
-}
-
-
 function openModal(boardId: string): void {
   showModal.value = true;
   newName.value = data.value!.find(board => board.whiteBoardId === boardId)?.title || ''
@@ -60,16 +45,6 @@ function openModal(boardId: string): void {
 function closeModal(): void {
   showModal.value = false;
 }
-
-function openNewBoardModal(): void {
-  showNewBoardModal.value = true;
-}
-
-function closeNewBoardModal(): void {
-  showNewBoardModal.value = false;
-  newBoardName.value = false
-}
-
 
 function logout() {
   userStore.logout()
@@ -166,10 +141,12 @@ function onEnter() {
 
           <div class="w-full md:w-1/3 px-3 mb-6 md:mb-0">
             <div class="border border-black border-t-1 border-b-4 border-x-1 p-3 rounded flex flex-col justify-content-flex-end hover:bg-gray-300">
+            <RouterLink to="/new">  
               <div class="flex items-center justify-center mb-2 cursor-pointer" @click="openNewBoardModal">
                 <PlusIcon/>
               </div>
               <h2 class="text-sm font-italic text-black text-center">Nueva pizarra</h2>
+            </RouterLink>
             </div>
           </div>
 
@@ -206,29 +183,7 @@ function onEnter() {
               </div>
             </div>
           </transition>
-
-
-          <transition>
-            <div class="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50 z-50" v-if="showNewBoardModal">
-              <div class="bg-white p-6 border border-black rounded-xl shadow-lg relative">
-                <button class="text-xl font-bold absolute top-1 right-2" @click="closeNewBoardModal">&times;</button>
-                <h1 class="flex justify-center font-bold mb-4">Nombre de la pizarra nueva:</h1>
-                <input class="border border-black rounded m-2 w-full" type="text" v-model="newBoardName">
-                <h1 class="flex justify-center font-bold mb-4">Selecciona un espacio de trabajo para la nueva pizarra:</h1>
-                <div class="flex justify-center mb-4">
-                  <select class="border border-black rounded m-2" v-model="selectedWorkspace">
-                    <option v-for="workspace in workspaces" :key="workspace.id" :value="workspace.id">{{ workspace.name }}</option>
-                  </select>
-                </div>
-                <div class="flex justify-center">
-                  <RouterLink to="/new">
-                    <button class="border border-black p-2 rounded hover:bg-green-400" @click="createNewBoard">Crear Pizarra</button>
-                  </RouterLink>
-                </div>
-              </div>
-            </div>
-          </transition>
-  
+          
         </div>
       </div>     
     </div>
@@ -241,7 +196,7 @@ function onEnter() {
 </div>
 </template>
 
-<style>
+
